javascript1: consistently use single backticks for inline code
In at least the "Introduction to JavaScript" chapter, code keywords such
as `let`, `const` and `this` are not consistently rendered as inline
code due to missing backticks. Sometimes, triple backticks (which work)
are used even though the code is inline.

I also amended line 337 to consistently blockquote the output.

Hopefully this improves readability to some degree, particularly when
referring to keywords such as `this` which can "blend in" otherwise.