Small clarification in typescript.md
I found the examples of Readonly<T> slightly confusing to read as it wasn't immediately clear that `studentVersion1:ImmutableStudent` is a new variable called studentVersion1 and not some modification of the original. Changing the `as const` declaration to version 2 and the Readonly<T> declaration to version 3 shows these are new variables. 

Also added a space after the colon to be consistent with other variable declarations.

This change may require a new screenshot of the 'squiggly' to match these code changes.